# docs/contracts/api.openapi.yaml
openapi: 3.1.0
info:
  title: VideoCutlist API
  version: 1.0.0
  description: >-
    Deployment authentication applies to application endpoints. API responses carry
    a server-generated X-Request-ID; errors contain error.code, error.message, and
    matching error.requestId. Missing/invalid authentication is 401, rejected origins
    are 403 origin_forbidden, and unexpected persistence/runtime failures are safe
    500 responses. Existing 422 input-validation exceptions are retained. HEAD
    responses have no body. MCP JSON-RPC is a separate protocol outside this contract.
servers:
  - url: "{serverBaseUrl}/api/v1"
    variables:
      serverBaseUrl:
        default: http://localhost:8787
        description: >-
          Reachable VideoCutlist server base URL; transport provider
          independent.
security:
  - bearerAuth: []
  - {}
paths:
...
  /media:
    get:
      summary: List indexed media
      operationId: listMedia
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaPage"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/tree:
    get:
      summary: Browse indexed media folders and items
      operationId: browseMediaTree
      parameters:
        - name: folderId
          in: query
          schema:
            type: string
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Media folder page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FolderPage"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/status:
    get:
      summary: Read media-library status
      operationId: mediaLibraryStatus
      responses:
        "200":
          description: Safe media-library status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LibraryStatus"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/refresh:
    post:
      summary: Refresh the media index
      operationId: refreshMedia
      responses:
        "202":
          description: Durable library scan job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "403":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/{mediaId}:
    parameters:
      - $ref: "#/components/parameters/MediaId"
    get:
      summary: Read media metadata
      operationId: getMedia
      responses:
        "200":
          description: Media metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Media"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /media/{mediaId}/preview:
    parameters:
      - $ref: "#/components/parameters/MediaId"
      - name: centerMs
        in: query
        required: true
        schema:
          type: integer
          minimum: 0
      - name: beforeMs
        in: query
        schema:
          type: integer
          minimum: 0
      - name: afterMs
        in: query
        schema:
          type: integer
          minimum: 0
      - name: mute
        in: query
        schema:
          type: boolean
          default: false
    get:
      summary: Stream a normalized preview
      operationId: streamPreview
      responses:
        "200":
          description: Fragmented MP4 preview
          headers:
            X-Preview-Start:
              schema:
                type: integer
            X-Preview-Duration:
              schema:
                type: integer
            X-Preview-Offset:
              schema:
                type: integer
            X-Preview-Cache:
              schema:
                type: string
                enum:
                  - hit
                  - miss
            X-Request-ID:
              schema:
                type: string
          content:
            video/mp4:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    head:
      summary: Check for a cached preview
      operationId: previewStatus
      responses:
        "200":
          description: Complete cached preview exists
        "404":
          description: No complete cached preview
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects:
    get:
      summary: List saved projects
      operationId: listProjects
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Project page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectPage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      summary: Read a project
      operationId: getProject
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    put:
      summary: Save a project
      operationId: putProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "200":
          description: Saved project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /projects/{projectId}/detections:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Start media detection
      operationId: createDetection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DetectionInput"
      responses:
        "202":
          description: Detection queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionJob"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects/{projectId}/exports:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Create a durable export
      operationId: createExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "202":
          description: Batch export queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchExportSubmission"
        "404":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /batches:
    get:
      summary: List durable batches
      operationId: listBatches
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Recent durable batches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchPage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /batches/{batchId}:
    parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Get durable batch export progress
      operationId: getBatchProgress
      responses:
        "200":
          description: Batch progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    delete:
      summary: Cancel a batch export
      operationId: cancelBatch
      responses:
        "204":
          description: >-
            Cancellation is requested for each queued or running child; 204 does
            not mean every runner has exited or its resources are released.
            Terminal children remain unchanged.
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /batches/{batchId}/download:
    parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
          pattern: ^b_[A-Za-z0-9_-]{12,64}$
    get:
      summary: Download all completed browser outputs as an archive
      operationId: downloadBatchOutputs
      responses:
        "200":
          description: >-
            Authenticated ZIP archive of verified browser-download outputs
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /projects/{projectId}/exports/preflight:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    post:
      summary: Validate export selection without queueing
      operationId: preflightExport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportInput"
      responses:
        "200":
          description: Structured export findings and deterministic selection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportPreflight"
        "422":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /destinations:
    get:
      summary: List configured export destinations
      operationId: listDestinations
      responses:
        "200":
          description: Public destination metadata; roots are never returned
          content:
            application/json:
              schema:
                type: object
                required:
                  - destinations
                  - capabilities
                properties:
                  destinations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Destination"
                  capabilities:
                    $ref: "#/components/schemas/DestinationCapabilities"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /jobs/{jobId}:
    parameters:
      - $ref: "#/components/parameters/JobId"
    get:
      summary: Read a job
      operationId: getJob
      responses:
        "200":
          description: Job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
    delete:
      summary: Cancel a job
      operationId: cancelJob
      responses:
        "204":
          description: >-
            Cancellation is durably accepted for queued or running jobs, but 204
            does not mean the runner process or its resources have stopped.
            Already terminal jobs are unchanged.
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /jobs/{jobId}/retry:
    parameters:
      - $ref: "#/components/parameters/JobId"
    post:
      summary: Retry a failed export as a new job
      operationId: retryJob
      responses:
        "202":
          description: New retry batch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "429":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /jobs/{jobId}/outputs/{position}:
    parameters:
      - $ref: "#/components/parameters/JobId"
      - name: position
        in: path
        required: true
        schema:
          type: integer
          minimum: 0
          maximum: 99
    get:
      summary: Download one completed output
      operationId: downloadJobOutput
      responses:
        "200":
          description: >-
            Authenticated attachment; only persisted download outputs are eligible
          content:
            video/x-matroska:
              schema:
                type: string
                format: binary
            video/mp4:
              schema:
                type: string
                format: binary
            video/quicktime:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
  /settings/media/refresh:
    post:
      summary: Queue a media library refresh
      operationId: refreshSettingsMedia
      responses:
        "202":
          description: Durable library scan job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Used only when the server selects static bearer mode.
...
  responses:
    Error:
      description: Safe error; the request ID matches X-Request-ID
      headers:
        X-Request-ID: {schema: {type: string}}
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    MediaPage:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Media"
        nextCursor:
          type:
            - string
            - "null"
    LibraryStatus:
      type: object
      required:
        - state
        - message
      properties:
        state:
          type: string
          enum:
            - unconfigured
            - scanning
            - ready_empty
            - ready_with_media
            - failed
        message:
          type: string
...